Provide optional model checkpointing in aggregator
