+++ fedscale/cloud/config_parser.py
@@ -54,6 +54,7 @@ parser.add_argument('--blacklist_max_len', type=float, default=0.3)
 parser.add_argument('--embedding_file', type=str,
                     default='glove.840B.300d.txt')
 parser.add_argument('--input_shape', type=int, nargs='+', default=[1, 3, 28, 28])
+parser.add_argument('--save_checkpoint', type=bool, default=False)
 
 
 # The configuration of different hyper-parameters for training
